--- src/features/resume/apis/hooks/resume/useCreateResume.ts
@@ -1,19 +1,0 @@
-import { AxiosResponse, AxiosError } from 'axios';
-import { UseMutationOptions, useMutation, useQueryClient } from '@tanstack/react-query';
-
-import { QUERY_KEY } from '@/shared/constants/queryKey';
-
-import resumeApi from '../../resume';
-
-const useCreateResume = (options?: UseMutationOptions<AxiosResponse, AxiosError>) => {
-  const queryClient = useQueryClient();
-
-  return useMutation(resumeApi.post, {
-    ...options,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [QUERY_KEY.resumes] });
-    },
-  });
-};
-
-export default useCreateResume;
