+++ src/features/resume/components/Aside/Aside.tsx
@@ -6,8 +6,8 @@ import IconPencil from '@/components/Icon/IconPencil';
 import Resume from './Resume/Resume';
 import ResumeListContainer from './Resume/ResumeListContainer';
 
-import useGetResumes from '../../apis/hooks/resume/useGetResumes';
-import useCreateResume from '../../apis/hooks/resume/useCreateResume';
+import { useGetResumes } from '@/hooks/reactQuery/resume/query';
+import { useCreateResume } from '@/hooks/reactQuery/resume/mutation';
 
 const Aside = () => {
   const { mutate: createResume } = useCreateResume();
